modif 31 mai 2024

+++ libft/dl_lstadd_front.c
@@ -6,13 +6,13 @@
 /*   By: najeuneh <najeuneh@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 14:27:35 by najeuneh          #+#    #+#             */
-/*   Updated: 2024/05/29 15:37:26 by najeuneh         ###   ########.fr       */
+/*   Updated: 2024/05/30 16:54:29 by najeuneh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-void	dl_lstadd_front(t_stack *a, int nbr)
+void	dl_lstadd_front(t_stack *a, long long nbr)
 {
 	t_node	*new;
