+++ libft/dl_lstnew.c
@@ -6,13 +6,13 @@
 /*   By: najeuneh <najeuneh@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 14:18:33 by najeuneh          #+#    #+#             */
-/*   Updated: 2024/05/28 17:24:01 by najeuneh         ###   ########.fr       */
+/*   Updated: 2024/05/30 17:00:37 by najeuneh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-t_node	*dl_lstnew(int nbr)
+t_node	*dl_lstnew(long long nbr)
 {
 	t_node	*node;
 	node = malloc(sizeof(t_node));
